# p2p/pex: optimize addrbook.hash