@odeke-em 2021-05-31
Rather:

p2p/pex: reuse hash.Hasher per addrbook for speed